requirements.txt
loguru
boto3
pyspark
findspark
mysql-connector-python
numpy
pandas
scikit-learn
matplotlib
mlflow
imblearn
apache-airflow
aws
requests